split the large lib.rs file
