+++ src/lib.rs
@@ -228,8 +228,37 @@
 //! # }
 //! ```
 
+// Module declarations
+mod constants;
+mod core;
+mod error;
+mod events;
+mod metrics;
+mod observer;
+mod persistence;
+mod subscription;
+
+pub mod computed;
+
+#[cfg(test)]
+pub mod testing;
+
+// Re-exports
+pub use error::PropertyError;
+pub use events::PropertyEvent;
+pub use metrics::PropertyMetrics;
+pub use observer::{Observer, ObserverId};
+pub use persistence::PropertyPersistence;
+pub use subscription::Subscription;
+pub use computed::computed;
+
+// Internal use
+use crate::constants::{MAX_THREADS, MAX_OBSERVERS};
+use crate::core::InnerProperty;
+use crate::observer::ObserverRef;
+#[cfg(feature = "debug")]
+use crate::events::ChangeLog;
 use std::collections::HashMap;
-use std::fmt;
 use std::mem;
 use std::panic;
 use std::sync::{Arc, Mutex, RwLock};
@@ -247,479 +276,6 @@ use std::pin::Pin;
 
 #[cfg(feature = "async")]
 use std::task::{Context, Poll};
-
-/// Maximum number of background threads used for asynchronous observer notifications
-///
-/// This constant controls the degree of parallelism when using `set_async()` to notify
-/// observers. The observer list is divided into batches, with each batch running in
-/// its own background thread, up to this maximum number of threads.
-///
-/// # Rationale
-///
-/// - **Resource Control**: Prevents unbounded thread creation that could exhaust system resources
-/// - **Performance Balance**: Provides parallelism benefits without excessive context switching overhead  
-/// - **Scalability**: Ensures consistent behavior regardless of the number of observers
-/// - **System Responsiveness**: Limits thread contention on multi-core systems
-///
-/// # Implementation Details
-///
-/// When `set_async()` is called:
-/// 1. All observers are collected into a snapshot
-/// 2. Observers are divided into `MAX_THREADS` batches (or fewer if there are fewer observers)
-/// 3. Each batch executes in its own `thread::spawn()` call
-/// 4. Observers within each batch are executed sequentially
-///
-/// For example, with 100 observers and `MAX_THREADS = 4`:
-/// - Batch 1: Observers 1-25 (Thread 1)
-/// - Batch 2: Observers 26-50 (Thread 2)  
-/// - Batch 3: Observers 51-75 (Thread 3)
-/// - Batch 4: Observers 76-100 (Thread 4)
-///
-/// # Tuning Considerations
-///
-/// This value can be adjusted based on your application's needs:
-/// - **CPU-bound observers**: Higher values may improve throughput on multi-core systems
-/// - **I/O-bound observers**: Higher values can improve concurrency for network/disk operations
-/// - **Memory-constrained systems**: Lower values reduce thread overhead
-/// - **Real-time systems**: Lower values reduce scheduling unpredictability
-///
-/// # Thread Safety
-///
-/// This constant is used only during the batching calculation and does not affect
-/// the thread safety of the overall system.
-const MAX_THREADS: usize = 4;
-
-/// Maximum number of observers allowed per property instance
-///
-/// This limit prevents memory exhaustion from unbounded observer registration.
-/// Once this limit is reached, attempts to add more observers will fail with
-/// an `InvalidConfiguration` error.
-///
-/// # Rationale
-///
-/// - **Memory Protection**: Prevents unbounded memory growth from observer accumulation
-/// - **Resource Management**: Ensures predictable memory usage in long-running applications
-/// - **Early Detection**: Catches potential memory leaks from forgotten unsubscriptions
-/// - **System Stability**: Prevents out-of-memory conditions in constrained environments
-///
-/// # Tuning Considerations
-///
-/// This value can be adjusted based on your application's needs:
-/// - **High-frequency properties**: May need fewer observers to avoid notification overhead
-/// - **Low-frequency properties**: Can safely support more observers
-/// - **Memory-constrained systems**: Lower values prevent memory pressure
-/// - **Development/testing**: Higher values may be useful for comprehensive test coverage
-///
-/// # Default Value
-///
-/// The default of 10,000 observers provides generous headroom for most applications while
-/// still preventing pathological cases. In practice, most properties have fewer than 100
-/// observers.
-///
-/// # Example Scenarios
-///
-/// - **User sessions**: 1,000 concurrent users, each with 5 properties = ~5,000 observers
-/// - **IoT devices**: 5,000 devices, each with 1-2 observers = ~10,000 observers
-/// - **Monitoring system**: 100 metrics, each with 20 dashboard widgets = ~2,000 observers
-const MAX_OBSERVERS: usize = 10_000;
-
-/// Trait for implementing property value persistence
-///
-/// This trait allows custom persistence strategies for ObservableProperty values,
-/// enabling automatic save/load functionality to various storage backends like
-/// disk files, databases, cloud storage, etc.
-///
-/// # Examples
-///
-/// ```rust
-/// use observable_property::PropertyPersistence;
-/// use std::fs;
-///
-/// struct FilePersistence {
-///     path: String,
-/// }
-///
-/// impl PropertyPersistence for FilePersistence {
-///     type Value = String;
-///
-///     fn load(&self) -> Result<Self::Value, Box<dyn std::error::Error + Send + Sync>> {
-///         fs::read_to_string(&self.path)
-///             .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync>)
-///     }
-///
-///     fn save(&self, value: &Self::Value) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-///         fs::write(&self.path, value)
-///             .map_err(|e| Box::new(e) as Box<dyn std::error::Error + Send + Sync>)
-///     }
-/// }
-/// ```
-pub trait PropertyPersistence: Send + Sync + 'static {
-    /// The type of value being persisted
-    type Value: Clone + Send + Sync + 'static;
-
-    /// Load the value from persistent storage
-    ///
-    /// # Returns
-    ///
-    /// Returns the loaded value or an error if loading fails
-    fn load(&self) -> Result<Self::Value, Box<dyn std::error::Error + Send + Sync>>;
-
-    /// Save the value to persistent storage
-    ///
-    /// # Arguments
-    ///
-    /// * `value` - The value to persist
-    ///
-    /// # Returns
-    ///
-    /// Returns `Ok(())` if successful, or an error if saving fails
-    fn save(&self, value: &Self::Value) -> Result<(), Box<dyn std::error::Error + Send + Sync>>;
-}
-
-/// Errors that can occur when working with ObservableProperty
-///
-/// # Note on Lock Poisoning
-///
-/// This implementation uses **graceful degradation** for poisoned locks. When a lock
-/// is poisoned (typically due to a panic in an observer or another thread), the
-/// library automatically recovers the inner value using [`PoisonError::into_inner()`](std::sync::PoisonError::into_inner).
-///
-/// This means:
-/// - **All operations continue to work** even after a lock is poisoned
-/// - No `ReadLockError`, `WriteLockError`, or `PoisonedLock` errors will occur in practice
-/// - The system remains operational and observers continue to function
-///
-/// The error variants are kept for backward compatibility and potential future use cases,
-/// but with the current implementation, poisoned locks are transparent to users.
-///
-/// # Production Benefit
-///
-/// This approach ensures maximum availability and resilience in production systems where
-/// a misbehaving observer shouldn't bring down the entire property system.
-#[derive(Debug, Clone)]
-pub enum PropertyError {
-    /// Failed to acquire a read lock on the property
-    ///
-    /// **Note**: With graceful degradation, this error is unlikely to occur in practice
-    /// as poisoned locks are automatically recovered.
-    ReadLockError {
-        /// Context describing what operation was being attempted
-        context: String,
-    },
-    /// Failed to acquire a write lock on the property
-    ///
-    /// **Note**: With graceful degradation, this error is unlikely to occur in practice
-    /// as poisoned locks are automatically recovered.
-    WriteLockError {
-        /// Context describing what operation was being attempted
-        context: String,
-    },
-    /// Attempted to unsubscribe an observer that doesn't exist
-    ObserverNotFound {
-        /// The ID of the observer that wasn't found
-        id: usize,
-    },
-    /// The property's lock has been poisoned due to a panic in another thread
-    ///
-    /// **Note**: With graceful degradation, this error will not occur in practice
-    /// as the implementation automatically recovers from poisoned locks.
-    PoisonedLock,
-    /// An observer function encountered an error during execution
-    ObserverError {
-        /// Description of what went wrong
-        reason: String,
-    },
-    /// The thread pool for async notifications is exhausted
-    ThreadPoolExhausted,
-    /// Invalid configuration was provided
-    InvalidConfiguration {
-        /// Description of the invalid configuration
-        reason: String,
-    },
-    /// Failed to persist the property value
-    PersistenceError {
-        /// Description of the persistence error
-        reason: String,
-    },
-    /// Attempted to undo when no history is available
-    NoHistory {
-        /// Description of the issue
-        reason: String,
-    },
-    /// Validation failed for the provided value
-    ValidationError {
-        /// Description of why validation failed
-        reason: String,
-    },
-}
-
-impl fmt::Display for PropertyError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            PropertyError::ReadLockError { context } => {
-                write!(f, "Failed to acquire read lock: {}", context)
-            }
-            PropertyError::WriteLockError { context } => {
-                write!(f, "Failed to acquire write lock: {}", context)
-            }
-            PropertyError::ObserverNotFound { id } => {
-                write!(f, "Observer with ID {} not found", id)
-            }
-            PropertyError::PoisonedLock => {
-                write!(
-                    f,
-                    "Property is in a poisoned state due to a panic in another thread"
-                )
-            }
-            PropertyError::ObserverError { reason } => {
-                write!(f, "Observer execution failed: {}", reason)
-            }
-            PropertyError::ThreadPoolExhausted => {
-                write!(f, "Thread pool is exhausted and cannot spawn more observers")
-            }
-            PropertyError::InvalidConfiguration { reason } => {
-                write!(f, "Invalid configuration: {}", reason)
-            }
-            PropertyError::PersistenceError { reason } => {
-                write!(f, "Persistence failed: {}", reason)
-            }
-            PropertyError::NoHistory { reason } => {
-                write!(f, "No history available: {}", reason)
-            }
-            PropertyError::ValidationError { reason } => {
-                write!(f, "Validation failed: {}", reason)
-            }
-        }
-    }
-}
-
-impl std::error::Error for PropertyError {}
-
-/// Function type for observers that get called when property values change
-pub type Observer<T> = Arc<dyn Fn(&T, &T) + Send + Sync>;
-
-/// Represents either a strong or weak reference to an observer
-///
-/// This enum allows the property to manage both strongly-held observers (which keep
-/// the observer alive) and weakly-held observers (which allow automatic cleanup when
-/// the observer is dropped elsewhere).
-enum ObserverRef<T: Clone + Send + Sync + 'static> {
-    /// Strong reference to an observer (keeps the observer alive)
-    Strong(Observer<T>),
-    /// Weak reference to an observer (automatically cleaned up when dropped)
-    Weak(std::sync::Weak<dyn Fn(&T, &T) + Send + Sync>),
-}
-
-impl<T: Clone + Send + Sync + 'static> ObserverRef<T> {
-    /// Attempts to get a callable observer from this reference
-    ///
-    /// For strong references, always returns Some. For weak references, attempts
-    /// to upgrade and returns None if the observer has been dropped.
-    fn try_call(&self) -> Option<Observer<T>> {
-        match self {
-            ObserverRef::Strong(arc) => Some(arc.clone()),
-            ObserverRef::Weak(weak) => weak.upgrade(),
-        }
-    }
-}
-
-/// Unique identifier for registered observers
-pub type ObserverId = usize;
-
-/// Performance metrics for an observable property
-///
-/// This struct provides insight into property usage patterns and observer
-/// notification performance, useful for debugging and performance optimization.
-///
-/// # Examples
-///
-/// ```rust
-/// use observable_property::ObservableProperty;
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let property = ObservableProperty::new(0);
-///
-/// property.subscribe(Arc::new(|old, new| {
-///     println!("Value changed: {} -> {}", old, new);
-/// }))?;
-///
-/// property.set(42)?;
-/// property.set(100)?;
-///
-/// let metrics = property.get_metrics()?;
-/// println!("Total changes: {}", metrics.total_changes);
-/// println!("Observer calls: {}", metrics.observer_calls);
-/// println!("Avg notification time: {:?}", metrics.avg_notification_time);
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug, Clone)]
-pub struct PropertyMetrics {
-    /// Total number of times the property value has changed
-    pub total_changes: usize,
-    /// Total number of observer calls (notification events)
-    pub observer_calls: usize,
-    /// Average time taken to notify all observers per change
-    pub avg_notification_time: Duration,
-}
-
-/// A RAII guard for an observer subscription that automatically unsubscribes when dropped
-///
-/// This struct provides automatic cleanup for observer subscriptions using RAII (Resource
-/// Acquisition Is Initialization) pattern. When a `Subscription` goes out of scope, its
-/// `Drop` implementation automatically removes the associated observer from the property.
-///
-/// This eliminates the need for manual `unsubscribe()` calls and helps prevent resource
-/// leaks in scenarios where observers might otherwise be forgotten.
-///
-/// # Type Requirements
-///
-/// The generic type `T` must implement the same traits as `ObservableProperty<T>`:
-/// - `Clone`: Required for observer notifications
-/// - `Send`: Required for transferring between threads  
-/// - `Sync`: Required for concurrent access from multiple threads
-/// - `'static`: Required for observer callbacks that may outlive the original scope
-///
-/// # Examples
-///
-/// ## Basic RAII Subscription
-///
-/// ```rust
-/// use observable_property::ObservableProperty;
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let property = ObservableProperty::new(0);
-///
-/// {
-///     // Create subscription - observer is automatically registered
-///     let _subscription = property.subscribe_with_subscription(Arc::new(|old, new| {
-///         println!("Value changed: {} -> {}", old, new);
-///     }))?;
-///
-///     property.set(42)?; // Observer is called: "Value changed: 0 -> 42"
-///
-///     // When _subscription goes out of scope here, observer is automatically removed
-/// }
-///
-/// property.set(100)?; // No observer output - subscription was automatically cleaned up
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## Cross-Thread Subscription Management
-///
-/// ```rust
-/// use observable_property::ObservableProperty;
-/// use std::sync::Arc;
-/// use std::thread;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let property = Arc::new(ObservableProperty::new(0));
-/// let property_clone = property.clone();
-///
-/// // Create subscription in main thread
-/// let subscription = property.subscribe_with_subscription(Arc::new(|old, new| {
-///     println!("Observed: {} -> {}", old, new);
-/// }))?;
-///
-/// // Move subscription to another thread for cleanup
-/// let handle = thread::spawn(move || {
-///     // Subscription is still active here
-///     let _ = property_clone.set(42); // Will trigger observer
-///     
-///     // When subscription is dropped here (end of thread), observer is cleaned up
-///     drop(subscription);
-/// });
-///
-/// handle.join().unwrap();
-/// 
-/// // Observer is no longer active
-/// property.set(100)?; // No output
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## Conditional Scoped Subscriptions
-///
-/// ```rust
-/// use observable_property::ObservableProperty;
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let counter = ObservableProperty::new(0);
-/// let debug_mode = true;
-///
-/// if debug_mode {
-///     let _debug_subscription = counter.subscribe_with_subscription(Arc::new(|old, new| {
-///         println!("Debug: counter {} -> {}", old, new);
-///     }))?;
-///     
-///     counter.set(1)?; // Prints debug info
-///     counter.set(2)?; // Prints debug info
-///     
-///     // Debug subscription automatically cleaned up when exiting if block
-/// }
-///
-/// counter.set(3)?; // No debug output (subscription was cleaned up)
-/// # Ok(())
-/// # }
-/// ```
-///
-/// # Thread Safety
-///
-/// Like `ObservableProperty` itself, `Subscription` is thread-safe. It can be safely
-/// sent between threads and the automatic cleanup will work correctly even if the
-/// subscription is dropped from a different thread than where it was created.
-pub struct Subscription<T: Clone + Send + Sync + 'static> {
-    inner: Arc<RwLock<InnerProperty<T>>>,
-    id: ObserverId,
-}
-
-impl<T: Clone + Send + Sync + 'static> std::fmt::Debug for Subscription<T> {
-    /// Debug implementation that shows the subscription ID without exposing internals
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Subscription")
-            .field("id", &self.id)
-            .field("inner", &"[ObservableProperty]")
-            .finish()
-    }
-}
-
-impl<T: Clone + Send + Sync + 'static> Drop for Subscription<T> {
-    /// Automatically removes the associated observer when the subscription is dropped
-    ///
-    /// This implementation provides automatic cleanup by removing the observer
-    /// from the property's observer list when the `Subscription` goes out of scope.
-    ///
-    /// # Error Handling
-    ///
-    /// If the property's lock is poisoned or inaccessible during cleanup, the error
-    /// is silently ignored using the `let _ = ...` pattern. This is intentional
-    /// because:
-    /// 1. Drop implementations should not panic
-    /// 2. If the property is poisoned, it's likely unusable anyway
-    /// 3. There's no meaningful way to handle cleanup errors in a destructor
-    ///
-    /// # Thread Safety
-    ///
-    /// This method is safe to call from any thread, even if the subscription
-    /// was created on a different thread.
-    fn drop(&mut self) {
-        // Graceful degradation: always attempt to clean up, even from poisoned locks
-        match self.inner.write() {
-            Ok(mut guard) => {
-                guard.observers.remove(&self.id);
-            }
-            Err(poisoned) => {
-                // Recover from poisoned lock to ensure cleanup happens
-                let mut guard = poisoned.into_inner();
-                guard.observers.remove(&self.id);
-            }
-        }
-    }
-}
 
 /// A custom Stream trait for async iteration over values
 ///
@@ -791,65 +347,6 @@ where
     inner: Arc<RwLock<InnerProperty<T>>>,
     max_threads: usize,
     max_observers: usize,
-}
-
-struct InnerProperty<T>
-where
-    T: Clone + Send + Sync + 'static,
-{
-    value: T,
-    observers: HashMap<ObserverId, ObserverRef<T>>,
-    next_id: ObserverId,
-    history: Option<Vec<T>>,
-    history_size: usize,
-    // Metrics tracking
-    total_changes: usize,
-    observer_calls: usize,
-    notification_times: Vec<Duration>,
-    // Debug tracking
-    #[cfg(feature = "debug")]
-    debug_logging_enabled: bool,
-    #[cfg(feature = "debug")]
-    change_logs: Vec<ChangeLog>,
-    // Change coalescing
-    batch_depth: usize,
-    batch_initial_value: Option<T>,
-    // Custom equality function
-    eq_fn: Option<Arc<dyn Fn(&T, &T) -> bool + Send + Sync>>,
-    // Validator function
-    validator: Option<Arc<dyn Fn(&T) -> Result<(), String> + Send + Sync>>,
-    // Event sourcing
-    event_log: Option<Vec<PropertyEvent<T>>>,
-    event_log_size: usize,
-}
-
-/// Information about a property change with stack trace
-#[cfg(feature = "debug")]
-#[derive(Clone)]
-struct ChangeLog {
-    timestamp: Instant,
-    old_value_repr: String,
-    new_value_repr: String,
-    backtrace: String,
-    thread_id: String,
-}
-
-/// Represents a single change event in the property's history
-///
-/// This struct captures all details of a value change for event sourcing,
-/// enabling time-travel debugging, audit logs, and event replay capabilities.
-#[derive(Clone, Debug)]
-pub struct PropertyEvent<T: Clone> {
-    /// When the change occurred
-    pub timestamp: Instant,
-    /// The value before the change
-    pub old_value: T,
-    /// The value after the change
-    pub new_value: T,
-    /// Sequential event number (starts at 0 for first change)
-    pub event_number: usize,
-    /// Thread that triggered the change (for debugging)
-    pub thread_id: String,
 }
 
 impl<T: Clone + Send + Sync + 'static> ObservableProperty<T> {
@@ -6307,410 +5804,6 @@ impl<'de, T: Clone + Send + Sync + 'static + Deserialize<'de>> Deserialize<'de> 
     {
         let value = T::deserialize(deserializer)?;
         Ok(ObservableProperty::new(value))
-    }
-}
-
-/// Creates a computed property that automatically recomputes when any dependency changes
-///
-/// A computed property is a special type of `ObservableProperty` whose value is derived
-/// from one or more other observable properties (dependencies). When any dependency changes,
-/// the computed property automatically recalculates its value and notifies its own observers.
-///
-/// # Type Parameters
-///
-/// * `T` - The type of the dependency properties
-/// * `U` - The type of the computed property's value
-/// * `F` - The compute function type
-///
-/// # Arguments
-///
-/// * `dependencies` - A vector of `Arc<ObservableProperty<T>>` that this computed property depends on
-/// * `compute_fn` - A function that takes a slice of current dependency values and returns the computed value
-///
-/// # Returns
-///
-/// An `Arc<ObservableProperty<U>>` that will automatically update when any dependency changes.
-/// Returns an error if subscribing to dependencies fails.
-///
-/// # How It Works
-///
-/// 1. Creates a new `ObservableProperty<U>` with the initial computed value
-/// 2. Subscribes to each dependency property
-/// 3. When any dependency changes, recomputes the value using all current dependency values
-/// 4. Updates the computed property, which triggers its own observers
-///
-/// # Examples
-///
-/// ## Basic Math Computation
-///
-/// ```rust
-/// use observable_property::{ObservableProperty, computed};
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// // Create source properties
-/// let width = Arc::new(ObservableProperty::new(10));
-/// let height = Arc::new(ObservableProperty::new(5));
-///
-/// // Create computed property for area
-/// let area = computed(
-///     vec![width.clone(), height.clone()],
-///     |values| values[0] * values[1]
-/// )?;
-///
-/// // Initial computed value
-/// assert_eq!(area.get()?, 50);
-///
-/// // Change width - area updates automatically
-/// width.set(20)?;
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// assert_eq!(area.get()?, 100);
-///
-/// // Change height - area updates automatically
-/// height.set(8)?;
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// assert_eq!(area.get()?, 160);
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## String Concatenation
-///
-/// ```rust
-/// use observable_property::{ObservableProperty, computed};
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let first_name = Arc::new(ObservableProperty::new("John".to_string()));
-/// let last_name = Arc::new(ObservableProperty::new("Doe".to_string()));
-///
-/// let full_name = computed(
-///     vec![first_name.clone(), last_name.clone()],
-///     |values| format!("{} {}", values[0], values[1])
-/// )?;
-///
-/// assert_eq!(full_name.get()?, "John Doe");
-///
-/// first_name.set("Jane".to_string())?;
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// assert_eq!(full_name.get()?, "Jane Doe");
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## Total Price with Tax
-///
-/// ```rust
-/// use observable_property::{ObservableProperty, computed};
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let subtotal = Arc::new(ObservableProperty::new(100.0_f64));
-/// let tax_rate = Arc::new(ObservableProperty::new(0.08_f64)); // 8% tax
-///
-/// let total = computed(
-///     vec![subtotal.clone(), tax_rate.clone()],
-///     |values| values[0] * (1.0 + values[1])
-/// )?;
-///
-/// assert_eq!(total.get()?, 108.0);
-///
-/// subtotal.set(200.0)?;
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// assert_eq!(total.get()?, 216.0);
-///
-/// tax_rate.set(0.10)?; // Change to 10%
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// // Use approximate comparison due to floating point precision
-/// let result = total.get()?;
-/// assert!((result - 220.0).abs() < 0.0001);
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## Observing Computed Properties
-///
-/// Computed properties are themselves `ObservableProperty` instances, so you can subscribe to them:
-///
-/// ```rust
-/// use observable_property::{ObservableProperty, computed};
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let a = Arc::new(ObservableProperty::new(5));
-/// let b = Arc::new(ObservableProperty::new(10));
-///
-/// let sum = computed(
-///     vec![a.clone(), b.clone()],
-///     |values| values[0] + values[1]
-/// )?;
-///
-/// // Subscribe to changes in the computed property
-/// sum.subscribe(Arc::new(|old, new| {
-///     println!("Sum changed from {} to {}", old, new);
-/// }))?;
-///
-/// a.set(7)?; // Will print: "Sum changed from 15 to 17"
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// # Ok(())
-/// # }
-/// ```
-///
-/// ## Chaining Computed Properties
-///
-/// You can create computed properties that depend on other computed properties:
-///
-/// ```rust
-/// use observable_property::{ObservableProperty, computed};
-/// use std::sync::Arc;
-///
-/// # fn main() -> Result<(), observable_property::PropertyError> {
-/// let celsius = Arc::new(ObservableProperty::new(0.0));
-///
-/// // First computed property: Celsius to Fahrenheit
-/// let fahrenheit = computed(
-///     vec![celsius.clone()],
-///     |values| values[0] * 9.0 / 5.0 + 32.0
-/// )?;
-///
-/// // Second computed property: Fahrenheit to Kelvin
-/// let kelvin = computed(
-///     vec![fahrenheit.clone()],
-///     |values| (values[0] - 32.0) * 5.0 / 9.0 + 273.15
-/// )?;
-///
-/// assert_eq!(celsius.get()?, 0.0);
-/// assert_eq!(fahrenheit.get()?, 32.0);
-/// assert_eq!(kelvin.get()?, 273.15);
-///
-/// celsius.set(100.0)?;
-/// std::thread::sleep(std::time::Duration::from_millis(10));
-/// assert_eq!(fahrenheit.get()?, 212.0);
-/// assert_eq!(kelvin.get()?, 373.15);
-/// # Ok(())
-/// # }
-/// ```
-///
-/// # Thread Safety
-///
-/// Computed properties are fully thread-safe. Updates happen asynchronously in response to
-/// dependency changes, and proper synchronization ensures the computed value is always
-/// based on the current dependency values at the time of computation.
-///
-/// # Performance Considerations
-///
-/// - The compute function is called every time any dependency changes
-/// - For expensive computations, consider using `subscribe_debounced` or `subscribe_throttled`
-///   on the dependencies before computing
-/// - The computed property uses async notifications, so there may be a small delay between
-///   a dependency change and the computed value update
-pub fn computed<T, U, F>(
-    dependencies: Vec<Arc<ObservableProperty<T>>>,
-    compute_fn: F,
-) -> Result<Arc<ObservableProperty<U>>, PropertyError>
-where
-    T: Clone + Send + Sync + 'static,
-    U: Clone + Send + Sync + 'static,
-    F: Fn(&[T]) -> U + Send + Sync + 'static,
-{
-    // Collect initial values from all dependencies
-    let initial_values: Result<Vec<T>, PropertyError> = 
-        dependencies.iter().map(|dep| dep.get()).collect();
-    let initial_values = initial_values?;
-    
-    // Compute initial value
-    let initial_computed = compute_fn(&initial_values);
-    
-    // Create the computed property
-    let computed_property = Arc::new(ObservableProperty::new(initial_computed));
-    
-    // Wrap compute_fn in Arc for sharing across multiple subscriptions
-    let compute_fn = Arc::new(compute_fn);
-    
-    // Subscribe to each dependency
-    for dependency in dependencies.iter() {
-        let deps_clone = dependencies.clone();
-        let computed_clone = computed_property.clone();
-        let compute_fn_clone = compute_fn.clone();
-        
-        // Subscribe to this dependency
-        dependency.subscribe(Arc::new(move |_old, _new| {
-            // When any dependency changes, collect all current values
-            let current_values: Result<Vec<T>, PropertyError> = 
-                deps_clone.iter().map(|dep| dep.get()).collect();
-            
-            if let Ok(values) = current_values {
-                // Recompute the value
-                let new_computed = compute_fn_clone(&values);
-                
-                // Update the computed property
-                if let Err(e) = computed_clone.set(new_computed) {
-                    eprintln!("Error updating computed property: {}", e);
-                }
-            }
-        }))?;
-    }
-    
-    Ok(computed_property)
-}
-
-/// Test helper utilities for working with `ObservableProperty` in tests.
-///
-/// This module provides convenient functions for testing observable properties,
-/// including waiting for notifications and collecting changes.
-#[cfg(test)]
-pub mod testing {
-    use super::*;
-    use std::sync::mpsc::{channel, Receiver};
-    use std::time::Duration;
-
-    /// Blocks the current thread until the next notification from the property.
-    ///
-    /// This is useful for synchronizing test code with property changes.
-    /// Times out after 5 seconds to prevent tests from hanging indefinitely.
-    ///
-    /// # Panics
-    ///
-    /// Panics if:
-    /// - The subscription fails
-    /// - No notification is received within 5 seconds
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use observable_property::{ObservableProperty, testing::await_notification};
-    /// use std::thread;
-    ///
-    /// let property = ObservableProperty::new(0);
-    ///
-    /// thread::spawn({
-    ///     let prop = property.clone();
-    ///     move || {
-    ///         thread::sleep(std::time::Duration::from_millis(100));
-    ///         prop.set(42).unwrap();
-    ///     }
-    /// });
-    ///
-    /// await_notification(&property);
-    /// assert_eq!(property.get().unwrap(), 42);
-    /// ```
-    pub fn await_notification<T: Clone + Send + Sync + 'static>(
-        property: &ObservableProperty<T>,
-    ) {
-        let (tx, rx) = channel();
-        
-        let _subscription = property
-            .subscribe_with_subscription(Arc::new(move |_old, _new| {
-                let _ = tx.send(());
-            }))
-            .expect("Failed to subscribe for await_notification");
-        
-        rx.recv_timeout(Duration::from_secs(5))
-            .expect("Timeout waiting for notification");
-    }
-
-    /// Records all changes to a property for test assertions.
-    ///
-    /// Returns a `ChangeCollector` that captures old and new values
-    /// whenever the property changes.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use observable_property::{ObservableProperty, testing::collect_changes};
-    ///
-    /// let property = ObservableProperty::new(0);
-    /// let collector = collect_changes(&property);
-    ///
-    /// property.set(1).unwrap();
-    /// property.set(2).unwrap();
-    /// property.set(3).unwrap();
-    ///
-    /// let changes = collector.changes();
-    /// assert_eq!(changes.len(), 3);
-    /// assert_eq!(changes[0], (0, 1));
-    /// assert_eq!(changes[1], (1, 2));
-    /// assert_eq!(changes[2], (2, 3));
-    /// ```
-    pub fn collect_changes<T: Clone + Send + Sync + 'static>(
-        property: &ObservableProperty<T>,
-    ) -> ChangeCollector<T> {
-        ChangeCollector::new(property)
-    }
-
-    /// Collects changes to an observable property for testing purposes.
-    ///
-    /// This struct maintains a subscription to a property and records
-    /// all changes (old value, new value) in a thread-safe manner.
-    pub struct ChangeCollector<T: Clone + Send + Sync + 'static> {
-        receiver: Receiver<(T, T)>,
-        _subscription: Subscription<T>,
-    }
-
-    impl<T: Clone + Send + Sync + 'static> ChangeCollector<T> {
-        /// Creates a new `ChangeCollector` that subscribes to the given property.
-        fn new(property: &ObservableProperty<T>) -> Self {
-            let (tx, rx) = channel();
-            
-            let subscription = property
-                .subscribe_with_subscription(Arc::new(move |old, new| {
-                    let _ = tx.send((old.clone(), new.clone()));
-                }))
-                .expect("Failed to subscribe for collect_changes");
-            
-            ChangeCollector {
-                receiver: rx,
-                _subscription: subscription,
-            }
-        }
-
-        /// Returns all collected changes as a vector of (old_value, new_value) tuples.
-        ///
-        /// This method drains all available changes from the internal receiver.
-        pub fn changes(&self) -> Vec<(T, T)> {
-            let mut changes = Vec::new();
-            while let Ok(change) = self.receiver.try_recv() {
-                changes.push(change);
-            }
-            changes
-        }
-
-        /// Waits for at least `count` changes to occur, with a timeout.
-        ///
-        /// Returns all collected changes once the count is reached.
-        /// Times out after 5 seconds.
-        ///
-        /// # Panics
-        ///
-        /// Panics if the expected number of changes is not received within the timeout.
-        pub fn wait_for_changes(&self, count: usize) -> Vec<(T, T)> {
-            let mut changes = Vec::new();
-            let timeout = Duration::from_secs(5);
-            let start = std::time::Instant::now();
-            
-            while changes.len() < count {
-                if start.elapsed() > timeout {
-                    panic!(
-                        "Timeout waiting for {} changes, only received {}",
-                        count,
-                        changes.len()
-                    );
-                }
-                
-                match self.receiver.recv_timeout(Duration::from_millis(100)) {
-                    Ok(change) => changes.push(change),
-                    Err(std::sync::mpsc::RecvTimeoutError::Timeout) => continue,
-                    Err(e) => panic!("Error receiving change: {}", e),
-                }
-            }
-            
-            changes
-        }
-
-        /// Returns the total number of changes currently collected.
-        pub fn count(&self) -> usize {
-            self.changes().len()
-        }
     }
 }
